requirements_docs.txt
cartopy==0.21.1
cmocean==2.0
colorcet==3.0.1
geovista==0.2.0
imageio>=2.5.0
imageio-ffmpeg==0.4.7
ipygany==0.5.0
ipyvtklink==0.2.3
jupyter_sphinx==0.4.0
jupyterlab==3.5.0
lxml==4.9.1
matplotlib==3.6.2
meshio==5.3.4
mypy==0.982
mypy-extensions==0.4.3
numpydoc==1.5.0
osmnx==1.2.2
panel==0.14.1
param==1.12.2  # due to panel bug
pydata-sphinx-theme==0.11.0
pypandoc==1.10
pyqt5
pytest-sphinx==0.5.0
pythreejs==2.4.1
pyvistaqt
QtPy
scipy==1.9.3
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinx-copybutton==0.5.0
sphinx-gallery==0.11.1
sphinx-notfound-page==0.8.3
sphinx-tabs
sphinx_design==0.3.0
sphinxcontrib-websupport==1.2.4
sphinxcontrib.asciinema
streamtracer==2.0.1
sunkit-pyvista==0.1.0
trame==2.5.0
trame-client==2.9.4
trame-server==2.11.4
trame-vtk==2.5.4
trimesh==3.15.8
typed-ast==1.5.4
typing_extensions==4.4.0
